requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.0.0
motor==3.3.2
pymongo==4.6.0
PyPDF2==3.0.1
python-docx==1.1.0
transformers==4.35.2
torch==2.1.0
requests==2.31.0
pydantic==2.5.0
numpy<2.0.0


